Change the rank property of Point/Region/Dist/Array/DistArray/etc be of type Long instead of type Int.  With the change to Long being the default integral type, it really makes sense for rank to be a Long.
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@25842 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.compiler/src/polyglot/types/Types.java
@@ -956,7 +956,7 @@ public class Types {
 	    XTerm selfZeroBased = xt.translate(self, zeroBasedField);
 	    XTerm selfRail = xt.translate(self, railField);
 
-	    XLit rankLiteral = ConstraintManager.getConstraintSystem().makeLit(1, ts.Int());
+	    XLit rankLiteral = ConstraintManager.getConstraintSystem().makeLit(1L, ts.Long());
 	    c.addBinding(selfRank, rankLiteral);
 	    c.addBinding(selfRect, ConstraintManager.getConstraintSystem().xtrue());
 	    c.addBinding(selfZeroBased, ConstraintManager.getConstraintSystem().xtrue());
